For a fresh start I moved main to old_main and created new main according to latest instructions. Also added a .gitignore file. Let's go from here, yeah?
